Perform_Word_Count function
Added a function that given a string array performs the word count, and returns a list of tupples that store the word and the number of ocurrences. The tupple list is sorted from the highest number of ocurrences to the lowest. Tested the function to find out other functions needed changes. Now Get_Stop_Word_List and Get_Book_Word_List return all the words in the list in lowercase. And in Get_Book_Word_List I changed the way of removing of punctuation marks.